+++ packages/plugins/contracts/test/typegen/nestedParentImport/A.sol
@@ -4,9 +4,9 @@ pragma solidity ^0.8.0;
 import { B } from "./D.sol";
 
 contract A is B {
-    uint public number;
+    uint256 public number;
 
-    constructor(uint _number, bool _boolean) B(_boolean) {
+    constructor(uint256 _number, bool _boolean) B(_boolean) {
         number = _number;
     }
